+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 'use client'
 import React, { useRef }        from 'react';
 
-import TextSplashPage           from '../components/text_splash_page';
-import BookingTransferFormCard  from '~/components/booking_transfer_interface';
+import TextSplashPage           from '../components/page-content/text_splash_page';
+import BookingTransferFormCard  from '~/components/page-content/booking_transfer_interface';
 
 
 export default function HomePage() {
