+++ src/components/elements/logo_instantiator.tsx
